+++ dhis-2/dhis-api/src/test/java/org/hisp/dhis/util/PeriodCriteriaUtilsTest.java
@@ -34,9 +34,6 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import java.lang.reflect.Field;
-import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Method;
 import java.util.Date;
 import java.util.HashSet;
 import java.util.Set;
@@ -226,32 +223,6 @@ class PeriodCriteriaUtilsTest {
     // when
     // then
     assertFalse(PeriodCriteriaUtils.hasPeriod(criteria) && criteria.getDimension().isEmpty());
-  }
-
-  @Test
-  void testCriteriaHasEmptySetsByDefault() {
-    // given
-    EventsAnalyticsQueryCriteria criteria = getDefaultEventsAnalyticsQueryCriteria();
-    Class<?> targetClass = EventsAnalyticsQueryCriteria.class;
-    Field[] fields = targetClass.getDeclaredFields();
-
-    for (Field field : fields) {
-      String fieldName = field.getName();
-      Class<?> fieldType = field.getType();
-      if (fieldType == Set.class) {
-        Method method;
-        try {
-          // when
-          method =
-              targetClass.getMethod(
-                  "get" + fieldName.substring(0, 1).toUpperCase() + fieldName.substring(1));
-          // then
-          assertTrue(((Set<?>) method.invoke(criteria)).isEmpty());
-        } catch (NoSuchMethodException | InvocationTargetException | IllegalAccessException e) {
-          assert false;
-        }
-      }
-    }
   }
 
   private EventsAnalyticsQueryCriteria configureEventsAnalyticsQueryCriteriaWithPeriod(
